rename validation-schema file to reflect the types of schemas
